top runs component and moved logic over
