+++ client/src/components/CharacterPageComponents/TopRuns/index.js
@@ -1,0 +1,3 @@
+import TopRuns from "./TopRuns";
+
+export default TopRuns;
