+++ client/src/components/CharacterPageComponents/index.js
@@ -2,5 +2,6 @@ import CharacterInfo from "./CharacterInfo";
 import Gear from "./Gear"
 import Talents from "./Talents";
 import Progression from "./Progression";
+import TopRuns from "./TopRuns";
 
-export { CharacterInfo, Gear, Talents, Progression }
+export { CharacterInfo, Gear, Talents, Progression, TopRuns }
