+++ client/src/pages/CharacterPage.js
@@ -1,7 +1,6 @@
 import { useParams, useLocation, Link } from "react-router-dom"
 import { useEffect, useState } from "react";
-import { v4 as uuid } from 'uuid';
-import { CharacterInfo, Gear, Talents, Progression } from "../components/CharacterPageComponents";
+import { CharacterInfo, Gear, Talents, Progression, TopRuns } from "../components/CharacterPageComponents";
 
 const CharacterPage = () => {
     const [equipment, setEquipment] = useState(null)
@@ -13,7 +12,7 @@ const CharacterPage = () => {
     const params = useParams();
     const charName = params.name.toLowerCase()
 
-    const dungeonArr = ['The Necrotic Wake', 'Mists of Tirna Scithe', 'Halls of Atonement', 'Spires of Ascension', 'De Other Side', 'Plaguefall', 'Theater of Pain', 'Sanguine Depths', 'Tazavesh: Streets of Wonder', 'Tazavesh: So\'leah\'s Gambit'];
+
 
     // fetches character info, M+ info, Raid info form RIO
     useEffect(() => {
@@ -46,16 +45,7 @@ const CharacterPage = () => {
         }
     }, [setTalents])
 
-    const timeFormat = (value) => {
-        const date = new Date(value)
 
-        if (date.getSeconds() >= 10) {
-            return `${date.getMinutes()}:${date.getSeconds()}`
-        } else {
-            return `${date.getMinutes()}:0${date.getSeconds()}`
-        }
-
-    }
 
     return (
         <section>
@@ -64,49 +54,9 @@ const CharacterPage = () => {
             <Gear equipment={equipment} />
             <Talents equipment={equipment} talents={talents} />
             <Progression equipment={equipment} currentRaid={currentRaid} setCurrentRaid={setCurrentRaid} currentSeason={currentSeason} setCurrentSeason={setCurrentSeason} />
+            <TopRuns dungeons={dungeons} />
 
-            <div>
-                <table>
-                    <thead>
-                        <tr>
-                            <th>Dungeon</th>
-                            <th>Level</th>
-                            <th>Affixes</th>
-                            <th>Time</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {dungeons !== null && (
-                            dungeonArr.map((d) => (
-                                dungeons.filter(dungeon => dungeon.dungeon.includes(d) && dungeon.affixes[0].name.includes('Tyrannical')).map(item => (
-                                    <tr key={uuid()}>
-                                        <td>{item.dungeon}</td>
-                                        <td>{`+${item.mythic_level}`}</td>
-                                        <td>{item.affixes.map(affix => (
-                                            <a href="#" data-wowhead={affix.wowhead_url} key={uuid()}><img src={`https://wow.zamimg.com/images/wow/icons/medium/${affix.icon}.jpg`}></img></a>
-                                        ))}</td>
-                                        <td>{timeFormat(item.clear_time_ms)}</td>
-                                    </tr>
-                                ))
-                            ))
-                        )}
-                        {dungeons !== null && (
-                            dungeonArr.map((d) => (
-                                dungeons.filter(dungeon => dungeon.dungeon.includes(d) && dungeon.affixes[0].name.includes('Fortified')).map(item => (
-                                    <tr key={uuid()}>
-                                        <td>{item.dungeon}</td>
-                                        <td>{`+${item.mythic_level}`}</td>
-                                        <td>{item.affixes.map(affix => (
-                                            <a href="#" data-wowhead={affix.wowhead_url} key={uuid()}><img src={`https://wow.zamimg.com/images/wow/icons/medium/${affix.icon}.jpg`}></img></a>
-                                        ))}</td>
-                                        <td>{timeFormat(item.clear_time_ms)}</td>
-                                    </tr>
-                                ))
-                            ))
-                        )}
-                    </tbody>
-                </table>
-            </div>
+
         </section>
     )
 }
